add a posts page to the new site template
When generating a new bridgetown site, it is not immediately clear how to get to your posts. I attempted a few ways to navigate there, but since I do not have prior experience with Jekyll, I was a little lost. This commit adds a posts page so that feature newbies, like myself, can quickly see their posts, and not have to try and spend time figuring it out.