Update pagination documentation (#98)
It do not currently see a place in the docs that we call out how to set the amount of paginated items per page. I had to go look in the source code, so this addition adds a little blurb about the `per_page` key to the pagination docs to make it easier to find.