build: add matrix testing for Ruby (#155)
* build: add matrix testing for Ruby

I thought GitHub may have access to Ruby 3 but I am not too hopefully after reconsidering. Let's find out.

* fix: uses 'ruby/setup' action instead and split job apart

Apparently Github's action is actually likely to be depricated in favor of the core ruby action. I also split the install and test command a part in case there are bugs. If the sets are in separate versions, it will help draw boundaries.

* fix: use correct action syntax

* fix: syntax...again

* fix: this time there was an error in the docs

* Test 2.x Rubies only

Co-authored-by: Jared White <jared@jaredwhite.com>