feat: add `yarn clean` script to package.json
What:

Adds a new script to our main `package.json` template, as well as the Bridgetown website.

`yarn clean`

Much like the build and serve commands, this is simply a npm script wrapping a BridgetownRB command.

Once executed, Bridgetown will search for files in the following locations:

- ./output.
- ./.bridgetown-metadata.
- ./.bridgetown-cache.
- ./.bridgetown-webpack.

If there are files present, it will remove them. Essentially clearing your cache and all remenants of the last build.

If the directory is empty, Bridgetown will alert you with something like: `Cleaner: Nothing to do for ./.bridgetown-webpack`.

Why:

I find myself always adding this script because it is nice shorthand to the longer `bundle exec bridgetown clean` or `bin/bridgetown clean`.

The reason I find myself reaching for it a lot is because its a nice final line of defense to determine if something is really broken or something
is just funky with the cache. When it comes to bundlers and tools like snowpack, it is even more handy due to how they load files.

I also find it good practice to always run a clean when the deploy script is executed. Sometimes I like to mimic production  locally
to test file sizes and small optimizations. Also, it couldn't hurt.

My final hope by placing it in the `package.json` is that it will be easier to find for newcomers and they may think to reach for it
sooner than others, which could save them an hour of debugging only to realize iCloud tried to store a file in the cloud or Webpack is being funky.