Add syntax-theme configuration command
This change introduces a new bundled configuration `syntax-theme`
which sets a syntax highlighting theme for code snippets.

To specify _which_ theme to use, this change introduces the concept
of a configuration option which is set by defining the option after
the configuration, separated by a colon `:`. Eg.

`bridgetown configure syntax-theme:github` (`github` is the option)

We use the themes from the Rouge gem to output the required CSS.