feat: add rb files to the Tailwind content glob (#564)
Components, plugins, helpers, data files, etc. that are written in Ruby may contain Tailwind CSS classes that won't get generated by the JIT because they aren't in the content scope. This commit updates the Tailwind content glob to also include ruby files when searching for classes in the src folder.