Add gemspec metadata for all gems
In trying to get to the repository for bridgetown, I noticed that the
main gem didn't have metadata set up for RubyGems.org. This change adds
the same metadata from `bridgetown-core` to all four other gems in the
repository to make it easier to get to the source code and other areas.