Use GitHub Actions to deploy GitHub Pages
Use the new way to deploy Bridgetown websites to GitHub Pages using
GitHub Actions.

This makes use of the [actions/deploy-pages action][deploy] action to
deploy a build artifact. Which relies on
[actions/upload-pages-artifact][upload] action to upload the artifact
generated by `bridgetown deploy`.

It's technically still in public beta, but I'm using it myself and seems
to work well enough.

Closes #614

[deploy]: https://github.com/actions/deploy-pages
[upload]: https://github.com/actions/upload-pages-artifact