Resolve Rubocop offenses (#644)
* Resolve some Rubcop offenses

- Resolve empty hash layout cop
- Remove redundant string escape sequences
- Disable class length cop for Collection class

* Specify period character in some regexes

Previously these parts of the resulting regex would match
any single character. Since we're constructing a regex
from a string, escaping of the forward slash is needed.

* Correct Zeitwerk setup for tests

Seems that we need to add a #setup call
in the more recent version of zeitwerk.

https://github.com/fxn/zeitwerk/blob/4d88e4ffeef65bdedc5077deae59e729b25a6c34/CHANGELOG.md#265-6-november-2022

* Update tests to reflect 'main' branch defaults

Now that 'main' is our fallback, we need a test for
the repo default in Utils.default_github_branch_name -
done here using a stubbed response.

* Correct a locale test

Need to find the locale by string instead of
symbol in this case as the relevant part of the .yml
is an item in a list.