Fix passing headers and add passing parameters to HTTP DSL (#722)
* Fix passing headers to GET requests

The original implementation ends up sending a `?headers=` query
parameter, which is not the intent.

This change ensures that it sends the data as headers instead of query
parameters.

* Allow passing GET parameters in the DSL

This change allows you to pass parameters to the `#get` DSL method as
keyword arguments. This makes the DSL easier to work with because you no
longer have to write the query string by hand.