doc: GitLab Pages serving compressed assets (#768)
* doc: enable gzip & brotli compression for Gitlab Pages

* doc: add a link to Gitlab Pages serving compressed assets

* doc: add missing Brotli dependency to compress Gitlab Pages assets

Brotli is a generic-purpose lossless compression algorithm that
compresses data using a combination of a modern variant of the LZ77
algorithm, Huffman coding and 2nd order context modeling, with a
compression ratio comparable to the best currently available
general-purpose compression methods. It is similar in speed with deflate
but offers more dense compression.