Eager load to make DEFAULTS deterministic
Configuration::DEFAULTS uses `Dir.pwd` to determine the root directory
for the configuration. This means that lazy loading the file using
`autoload` makes the root path for configuration non-deterministic
depending on the directory location the first time you try to use
`Bridgetown::Configuration`.

This has a knock-on effect of making tests fail nondeterministically.

By eager loading the configuration file, we can reduce this
nondeterminism by ensuring the value is set the same way every time.