Lock Rubocop to latest version that passes (#779)
This change stops ignoring the Gemfile.lock and checks it into the
repository. This is a best practice to make contributing easier for
new (and current) contributors.

One large paper cut is to get all of your changes to pass locally then
have CI fail because of updates in the linting pipeline that aren't
compatible with — in particular — code that you did not touch.

By checking Gemfile.lock in, it ensures that all contributors are
working from the same versions of dependencies that the CI server is
running.

In addition, this change also locks Rubocop to the newest version that
relies on rubocop-ast <= 1.28.0. That version and 1.29.0 have a parsing
bug in them that won't parse one of the files in Bridgetown's source
code and error out. Also, newer versions fail due to new cops on source
code that currently exists in the repository.