Refactor front matter handling and extract behavior into loaders (#778)
* Extract front matter loading into defined classes

This change refactors the loading of front matter into a series of
classes. This is a first step toward making front matter loaders
pluggable to allow, for example, [TOML][1]- or [KDL][2]-based front
matter.

[1]: https://toml.io
[2]: https://kdl.dev/

* Remove unused constants

These were developed prior to FrontMatterImporter and are no longer used.

* Consolidate front matter code in a module

This change reorganizes the front matter-related code so it is fully
contained within the Bridgetown::FrontMatter module. That makes it
easier to find the functionality and standardizes the way that
Bridgetown refers to front matter.

It also adds deprecated constant proxies for backwards-compatibility.
It's unclear whether the moved modules were public API or not so the
proxies make it so updating will not break anyone's site without
warning.

* Remove unnecessary include

This is already done by the include of the importer so isn't necessary
at this point.