Make OutputBuffer more performant for 3.2+ YJIT (#723)
This change ports Jean Boussier's Rails change over to Bridgetown with a
few tweaks.

1. Using `instance_of?` against the constant is faster than the
   alternative `self.class == other.class` construct
2. Linear, comparable logic for `<<`
3. An expanded set of delegated methods to include `#empty?`, `#encode`,
   `#lines`, `#reverse`, `#strip`, and `#valid_encoding?` because these
   are all called in the test suite.

In the current test suite, this is a count of how many times each method
is called:

2794 empty?
  54 encode
  54 encoding
 381 lines
 127 reverse
 127 strip
  54 valid_encoding?

It's unclear to me which, if any, of these should be public API so I
kept all of the delegated methods as well as the ones Jean picked for
Rails.